refactor: move mapman utils into the mapman component
